required steps done

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/JdbcUserDao.java
@@ -93,7 +93,7 @@ public class JdbcUserDao implements UserDao {
         user.setId(rs.getInt("user_id"));
         user.setUsername(rs.getString("username"));
         user.setPassword(rs.getString("password_hash"));
-        user.setBalance(rs.getInt("balance"));
+        user.setBalance(rs.getDouble("balance"));
         user.setActivated(true);
         user.setAuthorities("USER");
         return user;
